Fix bthci_cmd to bthci_evt association
There was a bug where a disconnect message between a command and it's
corresponding status or command complete event mangled the association
between the pending/response events to the command. When an event
arrived, the code scanned backwards from this frame for a command that
corresponds to the opcode which could generate such an event. If a
command for which the pending_in_frame or response_in_frame members were
not set was encountered, and it was the last with this property,
it was assumed to be the corresponding command and it was marked as such.
However, when a disconnect event arrives after a command, the
corresponding pending/complete event shall never arrive, leaving those
members unset. Then, when a command with the same opcode arrives, so do
it's corresponding events a while later but the latter are associated to the
command preceding the disconnect event.

This was fixed by traversing all seen commands once a disconnect
event is detected and setting their response_in_frame and
request_in_frame to 0.